帖子查看详情，点赞评论

+++ src/main/java/com/example/rednote/model/po/LikePO.java
@@ -11,9 +11,9 @@ import lombok.Data;
 @TableName("`like`")
 public class LikePO {
     @TableId(type = IdType.AUTO)
-    private Long id;
-    private Long userId;
-    private Long postId;
+    private Integer id;
+    private Integer userId;
+    private Integer postId;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss" , timezone = "GMT+8")
     private DateTime createdTime;
 }
